requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
chardet==5.0.0
click==8.1.3
configparser==5.3.0
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docker-pycreds==0.4.0
doufo==0.0.4
dxl-core==0.1.7
dxl-learn==0.2.1
dxl-shape==0.1.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flatbuffers==22.11.23
fonttools==4.38.0
fs==2.4.16
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.50.0
h5py==3.7.0
hydra-core==1.2.0
importlib-metadata==5.1.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.2
jfs==0.1.3
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.1
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-service==2.4.0
ml-collections==0.1.1
multidict==6.0.2
multipledispatch==0.6.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
olefile==0.46
omegaconf==2.2.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==3.2.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
sktime==0.14.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
statsmodels==0.13.5
subprocess32==3.5.4
tables==3.7.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.17.0
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.13.0
torchaudio==0.13.0
torchinfo==1.7.1
torchmetrics==0.10.3
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing==3.7.4.3
wandb==0.13.10
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
